refactor: centralize StrEnum compat shim and move enums to dedicated modules
Add xorq.common.compat with a single StrEnum import that falls back to
the strenum backport, and extract enum classes from large modules into
lightweight leaf modules (catalog/enums, common/enums, ibis_yaml/enums,
expr/ml/enums, backends/snowflake/enums, cli_constants).

Because the new enum modules only depend on xorq.common.compat, every
consumer can import them at the top level — all historical deferred
imports of these enums have been lifted and redundant
`from __future__ import annotations` lines removed from the leaf
modules.
