+++ python/xorq/catalog/constants.py
@@ -1,16 +1,6 @@
+from __future__ import annotations
+
 from pathlib import Path
-
-
-try:
-    from enum import StrEnum
-except ImportError:
-    from strenum import StrEnum
-
-
-class CatalogInfix(StrEnum):
-    ALIAS = "aliases"
-    ENTRY = "entries"
-    METADATA = "metadata"
 
 
 METADATA_APPEND = ".metadata.yaml"
