+++ python/xorq/catalog/enums.py
@@ -1,0 +1,20 @@
+from xorq.common.compat import StrEnum
+
+
+class CatalogInfix(StrEnum):
+    ALIAS = "aliases"
+    ENTRY = "entries"
+    METADATA = "metadata"
+
+
+class CatalogTag(StrEnum):
+    SOURCE = "catalog-source"
+    TRANSFORM = "catalog-transform"
+    CODE = "catalog-code"
+
+
+class OnUnrebuiltBuilder(StrEnum):
+    """Policy when a builder tag has no rebuild protocol registered."""
+
+    RAISE = "raise"
+    WARN = "warn"
