+++ python/xorq/common/compat.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+
+try:
+    from enum import StrEnum
+except ImportError:
+    from strenum import StrEnum
+
+__all__ = ["StrEnum"]
