+++ python/xorq/ibis_yaml/enums.py
@@ -1,7 +1,4 @@
-try:
-    from enum import StrEnum
-except ImportError:
-    from strenum import StrEnum
+from xorq.common.compat import StrEnum
 
 
 class DumpFiles(StrEnum):
@@ -34,3 +31,15 @@ class ExprKind(StrEnum):
 class MemtableTypes(StrEnum):
     inmemory = "memtables"
     database_table = "database_tables"
+
+
+class RefEnum(StrEnum):
+    dtype_ref = "dtype_ref"
+    node_ref = "node_ref"
+    schema_ref = "schema_ref"
+
+
+class RegistryEnum(StrEnum):
+    dtypes = "dtypes"
+    nodes = "nodes"
+    schemas = "schemas"
